requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
kaggle==1.6.14
python-dotenv==1.0.1
SQLAlchemy==2.0.30
